feat: divide settings into different pages
