+++ admin/SettingsPage.php
@@ -1,0 +1,8 @@
+<?php
+interface SettingsPage
+{
+    public function __construct($plugin);
+    public function renderer();
+    public function initFields();
+    public function initPage();
+}
